requirements-frozen.txt
astroid==2.9.3
autopep8==1.6.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
DateTime==4.4
flake8==4.0.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pylint==2.12.2
python-dotenv==0.19.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
termcolor==1.1.0
tinydb==4.7.0
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.0
wrapt==1.13.3
zope.interface==5.4.0
